Completely rewritten the RsyncSSH threading.Thread class. This addresses issue #1. It is now possible to have a configuration * `/project/theme` * `/project/theme_test` And it will only push to one folder. Before syncing theme would sync both to `theme` and `theme_test`
The problem was file system and subfolder inclusion tests done via string and substring comparisons which has to fail at some point. This is still present in the Rsync threading.Thread but there it is does not cause harm as the values pushed to it do not allow for it.