Regards #4. It was previously possible to use '.' as a configured folder if just one root exists. Currently this can't be done easily. Those keys are used in two ways:  1) as lookup in settings for the destination  2) as folders mapped to the real filesystem Those two use cases clash. Resolving the dot allows the checks to pass but the lookup fails and leaving the lookup intact complicates the checks massively.
There is just one valid use case where using dot syntax yields a benefit: single root projects with versioned project files. Here dot syntax works while the actual project location folder name is irrelevant.
This is not the main focus here and it will be closed as 'wontfix' until there is a way to do is nicely.

Added a guard that forbids dot syntax for now.